Move getFullDream useEffect from DreamView into DreamCard for ease of testing
